Add files via upload
I updated the vision sensor configuration to match that of the "Test Vision 6 colored objects Tuple Example" which works more consistently. As of now, when the robot drops off the object it tends to hit it when turning around to the original path. Also, it returns one step too far on the path.